+++ full-date-formatter/index.cjs
@@ -1,0 +1,4 @@
+'use strict'
+
+exports = module.exports = require('../commonjs/FullDateFormatter.js').default
+exports['default'] = require('../commonjs/FullDateFormatter.js').default