+++ full-date-formatter/index.d.ts
@@ -1,0 +1,6 @@
+import { Locale, DateInput } from '../index.d.js'
+
+export default class FullDateFormatter {
+	constructor(locale: Locale | Locale[]);
+	format(date: DateInput): string;
+}