+++ index.js
@@ -1,6 +1,8 @@
 export { default } from './modules/TimeAgo.js'
 
-// I guess these two are deprecated.
+export { default as FullDateFormatter } from './modules/FullDateFormatter.js'
+
+// These two legacy exports are deprecated.
 export {
 	intlDateTimeFormatSupported as intlDateTimeFormatSupported,
 	intlDateTimeFormatSupportedLocale as intlDateTimeFormatSupportedLocale