+++ test/TimeAgo.js
@@ -1,2 +1,2 @@
-// This file exists so that `setupLocales.js` could work from outside the `commonjs` folder.
+// This file exists so that `addLocales.js` could work from outside the `commonjs` folder.
 export { default } from '../source/TimeAgo.js'