+++ test/addLocales.js
@@ -1,0 +1,21 @@
+import TimeAgo from './TimeAgo.js'
+
+import en from '../locale/en.json' with { type: 'json' }
+import ru from '../locale/ru.json' with { type: 'json' }
+import de from '../locale/de.json' with { type: 'json' }
+import fr from '../locale/fr.json' with { type: 'json' }
+import es from '../locale/es.json' with { type: 'json' }
+import it from '../locale/it.json' with { type: 'json' }
+import ko from '../locale/ko.json' with { type: 'json' }
+import ccp from '../locale/ccp.json' with { type: 'json' }
+import zh from '../locale/zh.json' with { type: 'json' }
+
+TimeAgo.addLocale(en)
+TimeAgo.addLocale(ru)
+TimeAgo.addLocale(de)
+TimeAgo.addLocale(fr)
+TimeAgo.addLocale(es)
+TimeAgo.addLocale(it)
+TimeAgo.addLocale(ko)
+TimeAgo.addLocale(ccp)
+TimeAgo.addLocale(zh)