+++ test/globals.js
@@ -1,0 +1,4 @@
+import { should, expect } from 'chai'
+
+should()
+global.expect = expect